Create Finance Liquor Sales Analysis
This project is part of the Workearly Full Training in Data Science Certificate and its main goal is to simulate the daily routine as a Data Analyst while carrying out the analysis of a project. 

A dataset that contains Liquor Sales in the state of Iowa in USA between 2012-2020 was given. 

The main questions are the following ones: 

- Find the most popular item per zipcode in the period between 2016-2019
- Define the percentage of sales per store in the period between 2016-2019.

The insights of the analysis need to be visualized in either a matplotlib format or in Tableau Public.